LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-position-borders-002-expected.txt


FAIL anchor-position-borders-002 assert_array_equals: expected property 0 to be 807 but got 711 (expected array [807, 57, 807, 57] got [711, 26, 742, 57])
FAIL anchor-position-borders-002 1 assert_array_equals: expected property 0 to be 801 but got 705 (expected array [801, 119, 801, 119] got [705, 88, 736, 119])
FAIL anchor-position-borders-002 2 assert_array_equals: expected property 0 to be 8 but got 705 (expected array [8, 208, 8, 208] got [705, 155, 736, 186])
FAIL anchor-position-borders-002 3 assert_array_equals: expected property 0 to be 801 but got 705 (expected array [801, 255, 801, 255] got [705, 224, 736, 255])
FAIL anchor-position-borders-002 4 assert_array_equals: expected property 0 to be 801 but got 699 (expected array [801, 349, 801, 349] got [699, 296, 730, 327])
